requirements.txt
asgiref==3.8.1
Django==5.1.3
django-modeltranslation==0.19.11
gunicorn==20.1.0
numpy==2.2.1
pillow==11.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyarrow==18.1.0
pylance==0.21.0
python-dotenv==1.0.1
setuptools==75.6.0
sqlparse==0.5.2
tzdata==2024.2
